requirements.txt
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
boto3==1.17.54
django-storages==1.11.1
PyYAML==5.4.1
Pillow==7.2.0
django-autoslug==1.9.8
Markdown==3.3.4
sentry-sdk==1.0.0
requests==2.25.1
mailchimp==2.0.10
celery==5.0.5
django-celery-beat==2.2.0
django-celery-results==2.0.1
redis==3.5.3
